response entity for food

+++ src/main/java/card/cardapio/dto/food/FoodRequestDTO.java
@@ -1,7 +1,10 @@
 package card.cardapio.dto.food;
+
+import java.math.BigDecimal;
+
 public record FoodRequestDTO(
         String title,
-        Integer price,
+        BigDecimal price,
         String image,
         String description,
         String categoria,
